+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.7.1
+
+- fix typebox code generation for JSON schemas which use property names like
+  (without the quotes) "@test", "1", or "some with spaces".
+
+[src](https://github.com/xddq/schema2typebox/pull/36)
+
 # 1.7.0
 
 - ensure generation of $id as schema option in resulting typebox code (#33)